create LlvmIr2Hll pass in llvmir2hll lib
