+++ src/bin2llvmir/providers/config.cpp
@@ -27,63 +27,25 @@ namespace bin2llvmir {
 //=============================================================================
 //
 
-Config Config::empty(llvm::Module* m)
+retdec::config::Config _emptyConfig;
+
+Config::Config(retdec::config::Config& c)
+		: _configDB(c)
 {
-	Config config;
-	config._module = m;
-	return config;
+
 }
 
-Config Config::fromFile(llvm::Module* m, const std::string& path)
+Config Config::empty(llvm::Module* m)
 {
-	Config config;
+	Config config(_emptyConfig);
 	config._module = m;
-	config._configPath = path;
-	if (!config._configPath.empty())
-	{
-		// Can throw an exception, but it is catched by bin2llvmirl.
-		config._configDB.readJsonFile(config._configPath);
-	}
-
-	for (auto& s : config.getConfig().structures)
-	{
-		llvm_utils::stringToLlvmType(m->getContext(), s.getLlvmIr());
-	}
-
-	// TODO: needed?
-	if (config.getConfig().tools.isPic32())
-	{
-		config.getConfig().architecture.setIsPic32();
-	}
-
 	return config;
 }
 
 Config Config::fromConfig(llvm::Module* m, retdec::config::Config& c)
 {
-	Config config;
+	Config config(c);
 	config._module = m;
-	config._configDB = std::move(c);
-
-	// TODO: needed?
-	if (config.getConfig().tools.isPic32())
-	{
-		config.getConfig().architecture.setIsPic32();
-	}
-
-	return config;
-}
-
-Config Config::fromJsonString(llvm::Module* m, const std::string& json)
-{
-	Config config;
-	config._module = m;
-	config._configDB.readJsonString(json);
-
-	for (auto& s : config.getConfig().structures)
-	{
-		llvm_utils::stringToLlvmType(m->getContext(), s.getLlvmIr());
-	}
 
 	// TODO: needed?
 	if (config.getConfig().tools.isPic32())
@@ -102,11 +64,7 @@ void Config::doFinalization()
 {
 	tagFunctionsWithUsedCryptoGlobals();
 
-	if (!_configPath.empty())
-	{
-		_configDB.generateJsonFile(_configPath);
-	}
-	else if (!_configDB.parameters.getOutputConfigFile().empty())
+	if (!_configDB.parameters.getOutputConfigFile().empty())
 	{
 		_configDB.generateJsonFile(_configDB.parameters.getOutputConfigFile());
 	}
@@ -773,20 +731,6 @@ std::map<llvm::Module*, Config> ConfigProvider::_module2config;
 Config* ConfigProvider::addConfig(llvm::Module* m, retdec::config::Config& c)
 {
 	auto p = _module2config.emplace(m, Config::fromConfig(m, c));
-	return &p.first->second;
-}
-
-Config* ConfigProvider::addConfigFile(llvm::Module* m, const std::string& path)
-{
-	auto p = _module2config.emplace(m, Config::fromFile(m, path));
-	return &p.first->second;
-}
-
-Config* ConfigProvider::addConfigJsonString(
-		llvm::Module* m,
-		const std::string& json)
-{
-	auto p = _module2config.emplace(m, Config::fromJsonString(m, json));
 	return &p.first->second;
 }
