+++ src/retdec/retdec.cpp
@@ -47,67 +47,16 @@
 #include "retdec/bin2llvmir/providers/asm_instruction.h"
 #include "retdec/bin2llvmir/providers/config.h"
 
-#include "retdec/llvmir2hll/analysis/alias_analysis/alias_analysis.h"
-#include "retdec/llvmir2hll/analysis/alias_analysis/alias_analysis_factory.h"
-#include "retdec/llvmir2hll/analysis/value_analysis.h"
-#include "retdec/llvmir2hll/config/configs/json_config.h"
-#include "retdec/llvmir2hll/evaluator/arithm_expr_evaluator.h"
-#include "retdec/llvmir2hll/evaluator/arithm_expr_evaluator_factory.h"
-#include "retdec/llvmir2hll/graphs/cfg/cfg_builders/non_recursive_cfg_builder.h"
-#include "retdec/llvmir2hll/graphs/cfg/cfg_writer.h"
-#include "retdec/llvmir2hll/graphs/cfg/cfg_writer_factory.h"
-#include "retdec/llvmir2hll/graphs/cg/cg_builder.h"
-#include "retdec/llvmir2hll/graphs/cg/cg_writer.h"
-#include "retdec/llvmir2hll/graphs/cg/cg_writer_factory.h"
-#include "retdec/llvmir2hll/hll/hll_writer.h"
-#include "retdec/llvmir2hll/hll/hll_writer_factory.h"
-#include "retdec/llvmir2hll/ir/function.h"
-#include "retdec/llvmir2hll/ir/module.h"
-#include "retdec/llvmir2hll/llvm/llvm_debug_info_obtainer.h"
-#include "retdec/llvmir2hll/llvm/llvm_intrinsic_converter.h"
-#include "retdec/llvmir2hll/llvm/llvmir2bir_converter.h"
-#include "retdec/llvmir2hll/obtainer/call_info_obtainer.h"
-#include "retdec/llvmir2hll/obtainer/call_info_obtainer_factory.h"
-#include "retdec/llvmir2hll/optimizer/optimizer_manager.h"
-#include "retdec/llvmir2hll/pattern/pattern_finder_factory.h"
-#include "retdec/llvmir2hll/pattern/pattern_finder_runner.h"
-#include "retdec/llvmir2hll/pattern/pattern_finder_runners/cli_pattern_finder_runner.h"
-#include "retdec/llvmir2hll/pattern/pattern_finder_runners/no_action_pattern_finder_runner.h"
-#include "retdec/llvmir2hll/semantics/semantics/compound_semantics_builder.h"
-#include "retdec/llvmir2hll/semantics/semantics/default_semantics.h"
-#include "retdec/llvmir2hll/semantics/semantics_factory.h"
-#include "retdec/llvmir2hll/support/const_symbol_converter.h"
-#include "retdec/llvmir2hll/support/debug.h"
-#include "retdec/llvmir2hll/support/expr_types_fixer.h"
-#include "retdec/llvmir2hll/support/funcs_with_prefix_remover.h"
-#include "retdec/llvmir2hll/support/library_funcs_remover.h"
-#include "retdec/llvmir2hll/support/unreachable_code_in_cfg_remover.h"
-#include "retdec/llvmir2hll/utils/ir.h"
-#include "retdec/llvmir2hll/utils/string.h"
-#include "retdec/llvmir2hll/validator/validator.h"
-#include "retdec/llvmir2hll/validator/validator_factory.h"
-#include "retdec/llvmir2hll/var_name_gen/var_name_gen_factory.h"
-#include "retdec/llvmir2hll/var_name_gen/var_name_gens/num_var_name_gen.h"
-#include "retdec/llvmir2hll/var_renamer/var_renamer.h"
-#include "retdec/llvmir2hll/var_renamer/var_renamer_factory.h"
-#include "retdec/llvm-support/diagnostics.h"
-#include "retdec/utils/container.h"
-#include "retdec/utils/conversion.h"
-#include "retdec/utils/memory.h"
-#include "retdec/utils/string.h"
+#include "retdec/llvmir2hll/llvmir2hll.h"
 
 #include "retdec/config/config.h"
 #include "retdec/llvm-support/diagnostics.h"
 #include "retdec/retdec/retdec.h"
 #include "retdec/utils/memory.h"
 
-using retdec::llvmir2hll::ShPtr;
-using retdec::utils::hasItem;
-using retdec::utils::joinStrings;
-using retdec::utils::limitSystemMemory;
-using retdec::utils::limitSystemMemoryToHalfOfTotalSystemMemory;
-using retdec::utils::split;
-using retdec::utils::strToNum;
+//==============================================================================
+// disassembler
+//==============================================================================
 
 /**
  * Create an empty input module.
@@ -373,7 +322,7 @@ LlvmModuleContextPair disassemble(
 }
 
 //==============================================================================
-// bin2llvmir
+// decompiler
 //==============================================================================
 
 /**
@@ -492,1043 +441,10 @@ static inline void addPass(
 	PM.add(P);
 }
 
-//==============================================================================
-// llvmir2hll
-//==============================================================================
-
-std::string TargetHLL = "c";
-std::string OutputFormat = "plain";
-bool Debug = true;
-std::string Semantics = "";
-std::string ConfigPath = "";
-bool EmitDebugComments = true;
-std::string EnabledOpts = "";
-std::string DisabledOpts = "";
-bool NoOpts = false;
-bool AggressiveOpts = false;
-bool NoVarRenaming = false;
-bool NoSymbolicNames = false;
-bool KeepAllBrackets = false;
-bool KeepLibraryFunctions = false;
-bool NoTimeVaryingInfo = false;
-bool NoCompoundOperators = false;
-bool ValidateModule = false;
-std::string FindPatterns = "";
-std::string AliasAnalysis = "simple";
-std::string VarNameGen = "fruit";
-std::string VarNameGenPrefix = "";
-std::string VarRenamer = "readable";
-bool EmitCFGs = false;
-std::string CFGWriter = "dot";
-bool EmitCG = false;
-std::string CGWriter = "dot";
-std::string CallInfoObtainer = "optim";
-std::string ArithmExprEvaluator = "c";
-std::string ForcedModuleName = "";
-bool StrictFPUSemantics = false;
-std::string OutputFilename;
-
-std::unique_ptr<ToolOutputFile> getOutputStream(
-		const retdec::config::Parameters& params)
-{
-	// Open the file.
-	std::error_code ec;
-	auto out = std::make_unique<ToolOutputFile>(
-			params.getOutputFile(),
-			ec,
-			sys::fs::F_None
-	);
-	if (ec)
-	{
-		errs() << ec.message() << '\n';
-		return {};
-	}
-	return out;
-}
-
-/**
-* @brief Returns a list of all supported objects by the given factory.
-*
-* @tparam FactoryType Type of the factory in whose objects we are interested in.
-*
-* The list is comma separated and has no beginning or trailing whitespace.
-*/
-template<typename FactoryType>
-std::string getListOfSupportedObjects()
-{
-	return joinStrings(FactoryType::getInstance().getRegisteredObjects());
-}
-
-/**
-* @brief Prints an error message concerning the situation when an unsupported
-*        object has been selected from the given factory.
-*
-* @param[in] typeOfObjectsSingular A human-readable description of the type of
-*                                  objects the factory provides. In the
-*                                  singular form, e.g. "HLL writer".
-* @param[in] typeOfObjectsPlural A human-readable description of the type of
-*                                objects the factory provides. In the plural
-*                                form, e.g. "HLL writers".
-*
-* @tparam FactoryType Type of the factory in whose objects we are interested in.
-*/
-template<typename FactoryType>
-void printErrorUnsupportedObject(
-		const std::string &typeOfObjectsSingular,
-		const std::string &typeOfObjectsPlural)
-{
-	std::string supportedObjects(getListOfSupportedObjects<FactoryType>());
-	if (!supportedObjects.empty())
-	{
-		llvm_support::printErrorMessage(
-				"Invalid name of the ",
-				typeOfObjectsSingular,
-				" (supported names are: ",
-				supportedObjects,
-				")."
-		);
-	}
-	else
-	{
-		llvm_support::printErrorMessage(
-				"There are no available ",
-				typeOfObjectsPlural,
-				". Please SHIT, recompile the backend and try it again."
-		);
-	}
-}
-
-/**
-* @brief This class is the main chunk of code that converts an LLVM
-*        module to the specified high-level language (HLL).
-*
-* The decompilation is composed of the following steps:
-* 1) LLVM instantiates Decompiler with the output stream, where the target
-*    code will be emitted.
-* 2) The function runOnModule() is called, which decompiles the given
-*    LLVM IR into BIR (backend IR).
-* 3) The resulting IR is then converted into the requested HLL at the end of
-*    runOnModule().
-*
-* The HLL is specified in `-target-hll` when running llvmir2hll. Debug comments
-* can be enabled by using the `-emit-debug-comments` parameter. For more
-* information, run llvmir2hll with `-help`.
-*/
-class Decompiler: public ModulePass
-{
-public:
-	explicit Decompiler(
-			raw_pwrite_stream &out,
-			const retdec::config::Parameters& params
-	);
-
-	virtual llvm::StringRef getPassName() const override
-	{
-		return "Decompiler";
-	}
-	virtual bool runOnModule(Module &m) override;
-
-public:
-	/// Class identification.
-	static char ID;
-
-private:
-	virtual void getAnalysisUsage(AnalysisUsage &au) const override
-	{
-		au.addRequired<llvm::LoopInfoWrapperPass>();
-		au.addRequired<llvm::ScalarEvolutionWrapperPass>();
-		au.setPreservesAll();
-	}
-
-	bool initialize(Module &m);
-	void createSemantics();
-	void createSemanticsFromParameter();
-	void createSemanticsFromLLVMIR();
-	bool loadConfig();
-	void saveConfig();
-	bool convertLLVMIRToBIR();
-	void removeLibraryFuncs();
-	void removeCodeUnreachableInCFG();
-	void removeFuncsPrefixedWith(const retdec::llvmir2hll::StringSet &prefixes);
-	void fixSignedUnsignedTypes();
-	void convertLLVMIntrinsicFunctions();
-	void obtainDebugInfo();
-	void initAliasAnalysis();
-	void runOptimizations();
-	void renameVariables();
-	void convertConstantsToSymbolicNames();
-	void validateResultingModule();
-	void findPatterns();
-	void emitCFGs();
-	void emitCG();
-	void emitTargetHLLCode();
-	void finalize();
-	void cleanup();
-
-	llvmir2hll::StringSet parseListOfOpts(
-			const std::string &opts) const;
-	std::string getTypeOfRunOptimizations() const;
-	llvmir2hll::StringVector getIdsOfPatternFindersToBeRun() const;
-	llvmir2hll::PatternFinderRunner::PatternFinders instantiatePatternFinders(
-		const llvmir2hll::StringVector &pfsIds);
-	ShPtr<llvmir2hll::PatternFinderRunner> instantiatePatternFinderRunner() const;
-	llvmir2hll::StringSet getPrefixesOfFuncsToBeRemoved() const;
-
-private:
-	/// Output stream into which the generated code will be emitted.
-	raw_pwrite_stream &out;
-
-	const config::Parameters& params;
-
-	/// The input LLVM module.
-	Module *llvmModule;
-
-	/// The resulting module in BIR.
-	ShPtr<llvmir2hll::Module> resModule;
-
-	/// The used semantics.
-	ShPtr<llvmir2hll::Semantics> semantics;
-
-	/// The used config.
-	ShPtr<llvmir2hll::Config> config;
-
-	/// The used HLL writer.
-	ShPtr<llvmir2hll::HLLWriter> hllWriter;
-
-	/// The used alias analysis.
-	ShPtr<llvmir2hll::AliasAnalysis> aliasAnalysis;
-
-	/// The used obtainer of information about function and function calls.
-	ShPtr<llvmir2hll::CallInfoObtainer> cio;
-
-	/// The used evaluator of arithmetical expressions.
-	ShPtr<llvmir2hll::ArithmExprEvaluator> arithmExprEvaluator;
-
-	/// The used generator of variable names.
-	ShPtr<llvmir2hll::VarNameGen> varNameGen;
-
-	/// The used renamer of variables.
-	ShPtr<llvmir2hll::VarRenamer> varRenamer;
-};
-
-// Static variables and constants initialization.
-char Decompiler::ID = 0;
-
-/**
-* @brief Constructs a new decompiler.
-*
-* @param[in] out Output stream into which the generated HLL code will be
-*                emitted.
-*/
-Decompiler::Decompiler(raw_pwrite_stream &out, const config::Parameters& params)
-		:
-		ModulePass(ID),
-		out(out),
-		params(params),
-		llvmModule(nullptr),
-		resModule(),
-		semantics(),
-		hllWriter(),
-		aliasAnalysis(),
-		cio(),
-		arithmExprEvaluator(),
-		varNameGen(),
-		varRenamer()
-{
-
-}
-
-bool Decompiler::runOnModule(Module &m)
-{
-	llvm_support::printPhase("initialization", Debug);
-
-	bool decompilationShouldContinue = initialize(m);
-	if (!decompilationShouldContinue)
-	{
-		return false;
-	}
-
-	llvm_support::printPhase("conversion of LLVM IR into BIR", Debug);
-	decompilationShouldContinue = convertLLVMIRToBIR();
-	if (!decompilationShouldContinue)
-	{
-		return false;
-	}
-
-	llvmir2hll::StringSet funcPrefixes(getPrefixesOfFuncsToBeRemoved());
-	llvm_support::printPhase(
-			"removing functions prefixed with ["
-			+ joinStrings(funcPrefixes) + "]",
-			Debug
-	);
-	removeFuncsPrefixedWith(funcPrefixes);
-
-	if (!KeepLibraryFunctions)
-	{
-		llvm_support::printPhase(
-				"removing functions from standard libraries",
-				Debug
-		);
-		removeLibraryFuncs();
-	}
-
-	// The following phase needs to be done right after the conversion because
-	// there may be code that is not reachable in a CFG. This happens because
-	// the conversion of LLVM IR to BIR is not perfect, so it may introduce
-	// unreachable code. This causes problems later during optimizations
-	// because the code exists in BIR, but not in a CFG.
-	llvm_support::printPhase(
-			"removing code that is not reachable in a CFG",
-			Debug
-	);
-	removeCodeUnreachableInCFG();
-
-	llvm_support::printPhase("signed/unsigned types fixing", Debug);
-	fixSignedUnsignedTypes();
-
-	llvm_support::printPhase(
-			"converting LLVM intrinsic functions to standard functions",
-			Debug
-	);
-	convertLLVMIntrinsicFunctions();
-
-	if (resModule->isDebugInfoAvailable())
-	{
-		llvm_support::printPhase("obtaining debug information", Debug);
-		obtainDebugInfo();
-	}
-
-	if (!NoOpts)
-	{
-		llvm_support::printPhase(
-				"alias analysis [" + aliasAnalysis->getId() + "]",
-				Debug
-		);
-		initAliasAnalysis();
-
-		llvm_support::printPhase(
-				"optimizations [" + getTypeOfRunOptimizations()	+ "]",
-				Debug
-		);
-		runOptimizations();
-	}
-
-	if (!NoVarRenaming)
-	{
-		llvm_support::printPhase(
-				"variable renaming [" + varRenamer->getId() + "]",
-				Debug
-		);
-		renameVariables();
-	}
-
-	if (!NoSymbolicNames)
-	{
-		llvm_support::printPhase(
-				"converting constants to symbolic names",
-				Debug
-		);
-		convertConstantsToSymbolicNames();
-	}
-
-	if (ValidateModule)
-	{
-		llvm_support::printPhase("module validation", Debug);
-		validateResultingModule();
-	}
-
-	if (!FindPatterns.empty())
-	{
-		llvm_support::printPhase("finding patterns", Debug);
-		findPatterns();
-	}
-
-	if (EmitCFGs)
-	{
-		llvm_support::printPhase("emission of control-flow graphs", Debug);
-		emitCFGs();
-	}
-
-	if (EmitCG)
-	{
-		llvm_support::printPhase("emission of a call graph", Debug);
-		emitCG();
-	}
-
-	llvm_support::printPhase(
-			"emission of the target code [" + hllWriter->getId() + "]",
-			Debug
-	);
-	emitTargetHLLCode();
-
-	llvm_support::printPhase("finalization", Debug);
-	finalize();
-
-	llvm_support::printPhase("cleanup", Debug);
-	cleanup();
-
-	return false;
-}
-
-/**
-* @brief Initializes all the needed private variables.
-*
-* @return @c true if the decompilation should continue (the initialization went
-*         OK), @c false otherwise.
-*/
-bool Decompiler::initialize(Module &m)
-{
-	llvmModule = &m;
-
-	// Instantiate the requested HLL writer and make sure it exists. We need to
-	// explicitly specify template parameters because raw_pwrite_stream has
-	// a private copy constructor, so it needs to be passed by reference.
-	llvm_support::printSubPhase(
-			"creating the used HLL writer [" + TargetHLL + "]",
-			Debug
-	);
-	hllWriter = llvmir2hll::HLLWriterFactory::getInstance().createObject<
-		raw_pwrite_stream &>(TargetHLL, out, OutputFormat);
-	if (!hllWriter)
-	{
-		printErrorUnsupportedObject<llvmir2hll::HLLWriterFactory>(
-				"target HLL", "target HLLs"
-		);
-		return false;
-	}
-
-	// Instantiate the requested alias analysis and make sure it exists.
-	llvm_support::printSubPhase(
-			"creating the used alias analysis [" + AliasAnalysis + "]",
-			Debug
-	);
-	aliasAnalysis = llvmir2hll::AliasAnalysisFactory::getInstance().createObject(
-		AliasAnalysis
-	);
-	if (!aliasAnalysis)
-	{
-		printErrorUnsupportedObject<llvmir2hll::AliasAnalysisFactory>(
-				"alias analysis", "alias analyses"
-		);
-		return false;
-	}
-
-	// Instantiate the requested obtainer of information about function
-	// calls and make sure it exists.
-	llvm_support::printSubPhase(
-			"creating the used call info obtainer [" + CallInfoObtainer + "]",
-			Debug
-	);
-	cio = llvmir2hll::CallInfoObtainerFactory::getInstance().createObject(
-		CallInfoObtainer
-	);
-	if (!cio)
-	{
-		printErrorUnsupportedObject<llvmir2hll::CallInfoObtainerFactory>(
-				"call info obtainer", "call info obtainers"
-		);
-		return false;
-	}
-
-	// Instantiate the requested evaluator of arithmetical expressions and make
-	// sure it exists.
-	llvm_support::printSubPhase(
-			"creating the used evaluator of arithmetical expressions [" +
-			ArithmExprEvaluator + "]",
-			Debug
-	);
-	auto& aeef = llvmir2hll::ArithmExprEvaluatorFactory::getInstance();
-	arithmExprEvaluator = aeef.createObject(
-		ArithmExprEvaluator);
-	if (!arithmExprEvaluator)
-	{
-		printErrorUnsupportedObject<llvmir2hll::ArithmExprEvaluatorFactory>(
-				"evaluator of arithmetical expressions",
-				"evaluators of arithmetical expressions"
-		);
-		return false;
-	}
-
-	// Instantiate the requested variable names generator and make sure it
-	// exists.
-	llvm_support::printSubPhase(
-			"creating the used variable names generator [" + VarNameGen + "]",
-			Debug
-	);
-	varNameGen = llvmir2hll::VarNameGenFactory::getInstance().createObject(
-		VarNameGen,
-		VarNameGenPrefix
-	);
-	if (!varNameGen)
-	{
-		printErrorUnsupportedObject<llvmir2hll::VarNameGenFactory>(
-				"variable names generator", "variable names generators"
-		);
-		return false;
-	}
-
-	// Instantiate the requested variable renamer and make sure it exists.
-	llvm_support::printSubPhase(
-			"creating the used variable renamer [" + VarRenamer + "]",
-			Debug
-	);
-	varRenamer = llvmir2hll::VarRenamerFactory::getInstance().createObject(
-			VarRenamer,
-			varNameGen,
-			true
-	);
-	if (!varRenamer)
-	{
-		printErrorUnsupportedObject<llvmir2hll::VarRenamerFactory>(
-				"renamer of variables", "renamers of variables"
-		);
-		return false;
-	}
-
-	createSemantics();
-
-	bool configLoaded = loadConfig();
-	if (!configLoaded)
-	{
-		return false;
-	}
-
-	// Everything went OK.
-	return true;
-}
-
-/**
-* @brief Creates the used semantics.
-*/
-void Decompiler::createSemantics()
-{
-	if (!Semantics.empty())
-	{
-		// The user has requested some concrete semantics, so use it.
-		createSemanticsFromParameter();
-	}
-	else
-	{
-		// The user didn't request any semantics, so create it based on the
-		// data in the input LLVM IR.
-		createSemanticsFromLLVMIR();
-	}
-}
-
-/**
-* @brief Creates the used semantics as requested by the user.
-*/
-void Decompiler::createSemanticsFromParameter()
-{
-	if (Semantics.empty() || Semantics == "-")
-	{
-		// Do no use any semantics.
-		llvm_support::printSubPhase(
-				"creating the used semantics [none]",
-				Debug
-		);
-		semantics = llvmir2hll::DefaultSemantics::create();
-	}
-	else
-	{
-		// Use the given semantics.
-		llvm_support::printSubPhase(
-				"creating the used semantics [" + Semantics + "]",
-				Debug
-		);
-		semantics = llvmir2hll::CompoundSemanticsBuilder::build(
-				split(Semantics, ',')
-		);
-	}
-}
-
-/**
-* @brief Creates the used semantics based on the data in the input LLVM IR.
-*/
-void Decompiler::createSemanticsFromLLVMIR()
-{
-	// Create a list of the semantics to be used.
-	// TODO Use some data from the input LLVM IR, like the used compiler.
-	std::string usedSemantics("libc,gcc-general,win-api");
-
-	// Use the list to create the semantics.
-	llvm_support::printSubPhase(
-			"creating the used semantics [" + usedSemantics + "]",
-			Debug
-	);
-	semantics = llvmir2hll::CompoundSemanticsBuilder::build(
-			split(usedSemantics, ',')
-	);
-}
-
-/**
-* @brief Loads a config for the module.
-*
-* @return @a true if the config was loaded successfully, @c false otherwise.
-*/
-bool Decompiler::loadConfig()
-{
-	// Currently, we always use the JSON config.
-	if (ConfigPath.empty())
-	{
-		llvm_support::printSubPhase("creating a new config", Debug);
-		config = llvmir2hll::JSONConfig::empty();
-		return true;
-	}
-
-	llvm_support::printSubPhase("loading the input config", Debug);
-	try
-	{
-		config = llvmir2hll::JSONConfig::fromFile(ConfigPath);
-		return true;
-	}
-	catch (const llvmir2hll::ConfigError &ex)
-	{
-		llvm_support::printErrorMessage(
-			"Loading of the config failed: " + ex.getMessage() + "."
-		);
-		return false;
-	}
-}
-
-/**
-* @brief Saves the config file.
-*/
-void Decompiler::saveConfig()
-{
-	if (!ConfigPath.empty())
-	{
-		config->saveTo(ConfigPath);
-	}
-}
-
-/**
-* @brief Convert the LLVM IR module into a BIR module using the instantiated
-*        converter.
-* @return @c True if decompilation should continue, @c False if something went
-*         wrong and decompilation should abort.
-*/
-bool Decompiler::convertLLVMIRToBIR()
-{
-	auto llvm2BIRConverter = llvmir2hll::LLVMIR2BIRConverter::create(this);
-	// Options
-	llvm2BIRConverter->setOptionStrictFPUSemantics(StrictFPUSemantics);
-
-	std::string moduleName = ForcedModuleName.empty()
-			? llvmModule->getModuleIdentifier()
-			: ForcedModuleName;
-	resModule = llvm2BIRConverter->convert(
-			llvmModule,
-			moduleName,
-			semantics,
-			config,
-			Debug
-	);
-
-	return true;
-}
-
-/**
-* @brief Removes defined functions which are from some standard library whose
-*        header file has to be included because of some function declarations.
-*/
-void Decompiler::removeLibraryFuncs()
-{
-	llvmir2hll::FuncVector removedFuncs(
-			llvmir2hll::LibraryFuncsRemover::removeFuncs(resModule)
-	);
-
-	if (Debug)
-	{
-		// Emit the functions that were turned into declarations. Before that,
-		// however, sort them by name to provide a more deterministic output.
-		llvmir2hll::sortByName(removedFuncs);
-		for (const auto &func : removedFuncs)
-		{
-			llvm_support::printSubPhase("removing " + func->getName() + "()");
-		}
-	}
-}
-
-/**
-* @brief Removes code from all the functions in the module that is unreachable
-*        in the CFG.
-*/
-void Decompiler::removeCodeUnreachableInCFG()
-{
-	llvmir2hll::UnreachableCodeInCFGRemover::removeCode(resModule);
-}
-
-/**
-* @brief Removes functions with the given prefix.
-*/
-void Decompiler::removeFuncsPrefixedWith(
-		const retdec::llvmir2hll::StringSet &prefixes)
-{
-	llvmir2hll::FuncsWithPrefixRemover::removeFuncs(resModule, prefixes);
-}
-
-/**
-* @brief Fixes signed and unsigned types in the resulting module.
-*/
-void Decompiler::fixSignedUnsignedTypes()
-{
-	llvmir2hll::ExprTypesFixer::fixTypes(resModule);
-}
-
-/**
-* @brief Converts LLVM intrinsic functions to functions from the standard
-*        library.
-*/
-void Decompiler::convertLLVMIntrinsicFunctions()
-{
-	llvmir2hll::LLVMIntrinsicConverter::convert(resModule);
-}
-
-/**
-* @brief When available, obtains debugging information.
-*/
-void Decompiler::obtainDebugInfo()
-{
-	llvmir2hll::LLVMDebugInfoObtainer::obtainVarNames(resModule);
-}
-
-/**
-* @brief Initializes the alias analysis.
-*/
-void Decompiler::initAliasAnalysis()
-{
-	aliasAnalysis->init(resModule);
-}
-
-/**
-* @brief Runs the optimizations over the resulting module.
-*/
-void Decompiler::runOptimizations()
-{
-	ShPtr<llvmir2hll::OptimizerManager> optManager(
-			new llvmir2hll::OptimizerManager(
-					parseListOfOpts(EnabledOpts),
-					parseListOfOpts(DisabledOpts),
-					hllWriter,
-					llvmir2hll::ValueAnalysis::create(aliasAnalysis, true),
-					cio,
-					arithmExprEvaluator,
-					AggressiveOpts,
-					Debug
-			)
-	);
-	optManager->optimize(resModule);
-}
-
-/**
-* @brief Renames variables in the resulting module by using the selected
-*        variable renamer.
-*/
-void Decompiler::renameVariables()
-{
-	varRenamer->renameVars(resModule);
-}
-
-/**
-* @brief Converts constants in function calls to symbolic names.
-*/
-void Decompiler::convertConstantsToSymbolicNames()
-{
-	llvmir2hll::ConstSymbolConverter::convert(resModule);
-}
-
-/**
-* @brief Validates the resulting module.
-*/
-void Decompiler::validateResultingModule()
-{
-	// Run all the registered validators over the resulting module, sorted by
-	// name.
-	llvmir2hll::StringVector regValidatorIDs(
-		llvmir2hll::ValidatorFactory::getInstance().getRegisteredObjects()
-	);
-	std::sort(regValidatorIDs.begin(), regValidatorIDs.end());
-	for (const auto &id : regValidatorIDs)
-	{
-		llvm_support::printSubPhase("running " + id + "Validator", Debug);
-		ShPtr<llvmir2hll::Validator> validator(
-				llvmir2hll::ValidatorFactory::getInstance().createObject(id)
-		);
-		validator->validate(resModule, true);
-	}
-}
-
-/**
-* @brief Finds patterns in the resulting module.
-*/
-void Decompiler::findPatterns()
-{
-	llvmir2hll::StringVector pfsIds(getIdsOfPatternFindersToBeRun());
-	llvmir2hll::PatternFinderRunner::PatternFinders pfs(
-			instantiatePatternFinders(pfsIds)
-	);
-	ShPtr<llvmir2hll::PatternFinderRunner> pfr(
-			instantiatePatternFinderRunner()
-	);
-	pfr->run(pfs, resModule);
-}
-
-/**
-* @brief Emits the target HLL code.
-*/
-void Decompiler::emitTargetHLLCode()
-{
-	hllWriter->setOptionEmitDebugComments(EmitDebugComments);
-	hllWriter->setOptionKeepAllBrackets(KeepAllBrackets);
-	hllWriter->setOptionEmitTimeVaryingInfo(!NoTimeVaryingInfo);
-	hllWriter->setOptionUseCompoundOperators(!NoCompoundOperators);
-	hllWriter->emitTargetCode(resModule);
-}
-
-/**
-* @brief Finalizes the run of the back-end part.
-*/
-void Decompiler::finalize()
-{
-	saveConfig();
-}
-
-/**
-* @brief Cleanup.
-*/
-void Decompiler::cleanup()
-{
-	// Nothing to do.
-
-	// Note: Do not remove this phase, even if there is nothing to do. The
-	// presence of this phase is needed for the analyzing scripts in
-	// scripts/decompiler_tests (it marks the very last phase of a successful
-	// decompilation).
-}
-
-/**
-* @brief Emits a control-flow graph (CFG) for each function in the resulting
-*        module.
-*/
-void Decompiler::emitCFGs()
-{
-	// Make sure that the requested CFG writer exists.
-	llvmir2hll::StringVector availCFGWriters(
-		llvmir2hll::CFGWriterFactory::getInstance().getRegisteredObjects());
-	if (!hasItem(availCFGWriters, std::string(CFGWriter)))
-	{
-		printErrorUnsupportedObject<llvmir2hll::CFGWriterFactory>(
-			"CFG writer", "CFG writers");
-		return;
-	}
-
-	// Instantiate a CFG builder.
-	ShPtr<llvmir2hll::CFGBuilder> cfgBuilder(
-			llvmir2hll::NonRecursiveCFGBuilder::create()
-	);
-
-	// Get the extension of the files that will be written (we use the CFG
-	// writer's name for this purpose).
-	std::string fileExt(CFGWriter);
-
-	// For each function in the resulting module...
-	for (auto i = resModule->func_definition_begin(),
-			e = resModule->func_definition_end();
-			i != e;
-			++i)
-	{
-		// Open the output file.
-		std::string fileName(
-				OutputFilename + ".cfg." + (*i)->getName() + "." + fileExt
-		);
-		std::ofstream out(fileName.c_str());
-		if (!out)
-		{
-			llvm_support::printErrorMessage(
-					"Cannot open " + fileName + " for writing."
-			);
-			return;
-		}
-		// Create a CFG for the current function and emit it into the opened
-		// file.
-		auto& cfgwf = llvmir2hll::CFGWriterFactory::getInstance();
-		ShPtr<llvmir2hll::CFGWriter> writer(
-				cfgwf.createObject<ShPtr<llvmir2hll::CFG>, std::ostream &>(
-						CFGWriter,
-						cfgBuilder->getCFG(*i),
-						out
-				)
-		);
-		ASSERT_MSG(
-				writer,
-				"instantiation of the requested CFG writer `"
-				<< CFGWriter << "` failed"
-		);
-		writer->emitCFG();
-	}
-}
-
-/**
-* @brief Emits a call graph (CG) for the resulting module.
-*/
-void Decompiler::emitCG()
-{
-	// Make sure that the requested CG writer exists.
-	auto& inst = llvmir2hll::CGWriterFactory::getInstance();
-	llvmir2hll::StringVector availCGWriters(
-			inst.getRegisteredObjects()
-	);
-	if (!hasItem(availCGWriters, std::string(CGWriter)))
-	{
-		printErrorUnsupportedObject<llvmir2hll::CGWriterFactory>(
-				"CG writer", "CG writers"
-		);
-		return;
-	}
-
-	// Get the extension of the file that will be written (we use the CG
-	// writer's name for this purpose).
-	std::string fileExt(CGWriter);
-
-	// Open the output file.
-	std::string fileName(OutputFilename + ".cg." + fileExt);
-	std::ofstream out(fileName.c_str());
-	if (!out)
-	{
-		llvm_support::printErrorMessage(
-				"Cannot open " + fileName + " for writing."
-		);
-		return;
-	}
-
-	// Create a CG for the current module and emit it into the opened file.
-	auto& cgwf = llvmir2hll::CGWriterFactory::getInstance();
-	ShPtr<llvmir2hll::CGWriter> writer(
-			cgwf.createObject<ShPtr<llvmir2hll::CG>, std::ostream &>(
-			CGWriter, llvmir2hll::CGBuilder::getCG(resModule), out
-	));
-	ASSERT_MSG(
-			writer,
-			"instantiation of the requested CG writer `"
-			<< CGWriter << "` failed"
-	);
-	writer->emitCG();
-}
-
-/**
-* @brief Parses the given list of optimizations.
-*
-* @a opts should be a list of strings separated by a comma.
-*/
-retdec::llvmir2hll::StringSet Decompiler::parseListOfOpts(
-		const std::string &opts) const
-{
-	llvmir2hll::StringVector parsedOpts(split(opts, ','));
-	return llvmir2hll::StringSet(parsedOpts.begin(), parsedOpts.end());
-}
-
-/**
-* @brief Returns the type of optimizations that should be run (as a string).
-*/
-std::string Decompiler::getTypeOfRunOptimizations() const
-{
-	return AggressiveOpts ? "aggressive" : "normal";
-}
-
-/**
-* @brief Returns the IDs of pattern finders to be run.
-*/
-retdec::llvmir2hll::StringVector
-Decompiler::getIdsOfPatternFindersToBeRun() const
-{
-	if (FindPatterns == "all")
-	{
-		// Get all of them.
-		auto& inst = llvmir2hll::PatternFinderFactory::getInstance();
-		return inst.getRegisteredObjects();
-	}
-	else
-	{
-		// Get only the selected IDs.
-		return split(FindPatterns, ',');
-	}
-}
-
-/**
-* @brief Instantiates and returns the pattern finders described by their ID.
-*
-* If a pattern finder cannot be instantiated, a warning message is emitted.
-*/
-retdec::llvmir2hll::PatternFinderRunner::PatternFinders
-Decompiler::instantiatePatternFinders(
-		const retdec::llvmir2hll::StringVector &pfsIds)
-{
-	// Pattern finders need a value analysis, so create it.
-	initAliasAnalysis();
-	ShPtr<llvmir2hll::ValueAnalysis> va(
-			llvmir2hll::ValueAnalysis::create(aliasAnalysis, true));
-
-	// Re-initialize cio to be sure its up-to-date.
-	cio->init(llvmir2hll::CGBuilder::getCG(resModule), va);
-
-	llvmir2hll::PatternFinderRunner::PatternFinders pfs;
-	for (const auto pfId : pfsIds)
-	{
-		auto& inst = llvmir2hll::PatternFinderFactory::getInstance();
-		ShPtr<llvmir2hll::PatternFinder> pf(
-				inst.createObject(pfId, va, cio)
-		);
-		if (!pf && Debug)
-		{
-			llvm_support::printWarningMessage(
-				"the requested pattern finder '" + pfId + "' does not exist");
-		}
-		else
-		{
-			pfs.push_back(pf);
-		}
-	}
-	return pfs;
-}
-
-/**
-* @brief Instantiates and returns a proper PatternFinderRunner.
-*/
-ShPtr<retdec::llvmir2hll::PatternFinderRunner>
-Decompiler::instantiatePatternFinderRunner() const
-{
-	if (Debug)
-	{
-		return ShPtr<llvmir2hll::PatternFinderRunner>(
-			new llvmir2hll::CLIPatternFinderRunner(llvm::errs()));
-	}
-	return ShPtr<llvmir2hll::PatternFinderRunner>(
-		new llvmir2hll::NoActionPatternFinderRunner()
-	);
-}
-
-/**
-* @brief Returns the prefixes of functions to be removed.
-*/
-retdec::llvmir2hll::StringSet Decompiler::getPrefixesOfFuncsToBeRemoved() const
-{
-	return config->getPrefixesOfFuncsToBeRemoved();
-}
-
-//==============================================================================
-// bin2llvmir + llvmir2hll
-//==============================================================================
-
 llvm::Target decompilerTarget;
 
 bool decompile(const retdec::config::Parameters& params)
 {
-
-//==============================================================================
-// bin2llvmir
-//==============================================================================
-
 	llvm_support::printPhase("Initialization");
 	auto& passRegistry = initializeLlvmPasses();
 
@@ -1570,15 +486,12 @@ c.parameters.staticSignaturePaths =
 	"/home/peter/retdec/retdec/build/install/share/retdec/support/generic/yara_patterns/static-code",
 };
 
-ConfigPath = c.parameters.getOutputConfigFile();
-OutputFilename = c.parameters.getOutputFile();
-
-pm.add(new bin2llvmir::ProviderInitialization(&c));
+// pm.add(new bin2llvmir::ProviderInitialization(&c));
 
 	std::vector<std::string> passes =
 	{
 		// retdec
-		//"provider-init",
+		"provider-init",
 		"decoder",
 		"verify",
 		"x86-addr-spaces",
@@ -1716,7 +629,7 @@ pm.add(new bin2llvmir::ProviderInitialization(&c));
 		"idioms",
 		"remove-phi",
 		"value-protect",
-		"write-config",
+//		"write-config",
 		"sink",
 		"verify",
 		"write-ll",
@@ -1724,43 +637,34 @@ pm.add(new bin2llvmir::ProviderInitialization(&c));
 		// llvmir2hll
 		"loops",
 		"scalar-evolution",
+		"llvmir2hll",
 	};
 	for (auto& p : passes)
 	{
 		if (auto* info = passRegistry.getPassInfo(p))
 		{
-			if (auto* ctr = info->getNormalCtor())
+			auto* pass = info->createPass();
+			addPass(pm, pass);
+
+			if (info->getTypeInfo() == &bin2llvmir::ProviderInitialization::ID)
 			{
-				addPass(pm, ctr());
-				continue;
+				auto* p = static_cast<bin2llvmir::ProviderInitialization*>(pass);
+				p->setConfig(&c);
+			}
+			if (info->getTypeInfo() == &llvmir2hll::LlvmIr2Hll::ID)
+			{
+				auto* p = static_cast<llvmir2hll::LlvmIr2Hll*>(pass);
+				p->setConfig(&c);
 			}
 		}
-
-		throw std::runtime_error("cannot create pass: " + p);
+		else
+		{
+			throw std::runtime_error("cannot create pass: " + p);
+		}
 	}
-
-//==============================================================================
-// llvmir2hll
-//==============================================================================
-
-	// Figure out where we are going to send the output.
-	auto out = getOutputStream(params);
-	if (!out) {
-		return EXIT_FAILURE;
-	}
-
-	raw_pwrite_stream &os(out->os());
-	pm.add(new Decompiler(os, params));
-
-//==============================================================================
-// together
-//==============================================================================
 
 	// Now that we have all of the passes ready, run them.
 	pm.run(*module);
-
-	// Declare success.
-	out->keep();
 
 	return EXIT_SUCCESS;
 }
