+++ tests/bin2llvmir/optimizations/param_return/param_return_tests.cpp
@@ -46,7 +46,7 @@ TEST_F(ParamReturnTests, x86PtrCallBasicFunctionality)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 32,
 			"endian" : "little",
@@ -69,6 +69,7 @@ TEST_F(ParamReturnTests, x86PtrCallBasicFunctionality)
 			}
 		]
 	})");
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 	auto typeConfig = std::make_unique<ctypesparser::TypeConfig>();
 	auto demangler = DemanglerProvider::addDemangler(
@@ -113,7 +114,7 @@ TEST_F(ParamReturnTests, x86PtrCallPrevBbIsUsedOnlyIfItIsASinglePredecessor)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 32,
 			"endian" : "little",
@@ -136,6 +137,7 @@ TEST_F(ParamReturnTests, x86PtrCallPrevBbIsUsedOnlyIfItIsASinglePredecessor)
 			}
 		]
 	})");
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 	auto typeConfig = std::make_unique<ctypesparser::TypeConfig>();
 	auto demangler = DemanglerProvider::addDemangler(
@@ -185,7 +187,7 @@ TEST_F(ParamReturnTests, x86PtrCallPrevBbIsNotUsedIfItIsNotASinglePredecessor)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 32,
 			"endian" : "little",
@@ -208,6 +210,7 @@ TEST_F(ParamReturnTests, x86PtrCallPrevBbIsNotUsedIfItIsNotASinglePredecessor)
 			}
 		]
 	})");
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 	auto typeConfig = std::make_unique<ctypesparser::TypeConfig>();
 	auto demangler = DemanglerProvider::addDemangler(
@@ -254,7 +257,7 @@ TEST_F(ParamReturnTests, x86PtrCallOnlyStackStoresAreUsed)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 32,
 			"endian" : "little",
@@ -273,6 +276,7 @@ TEST_F(ParamReturnTests, x86PtrCallOnlyStackStoresAreUsed)
 			}
 		]
 	})");
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 
 	abi->addRegister(X86_REG_EAX, getGlobalByName("eax"));
@@ -320,7 +324,7 @@ TEST_F(ParamReturnTests, x86PtrCallStackAreUsedAsArgumentsInCorrectOrder)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 32,
 			"endian" : "little",
@@ -343,6 +347,7 @@ TEST_F(ParamReturnTests, x86PtrCallStackAreUsedAsArgumentsInCorrectOrder)
 			}
 		]
 	})");
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 	auto typeConfig = std::make_unique<ctypesparser::TypeConfig>();
 	auto demangler = DemanglerProvider::addDemangler(
@@ -387,7 +392,7 @@ TEST_F(ParamReturnTests, x86PtrCallOnlyContinuousStackOffsetsAreUsed)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 32,
 			"endian" : "little",
@@ -418,6 +423,7 @@ TEST_F(ParamReturnTests, x86PtrCallOnlyContinuousStackOffsetsAreUsed)
 			}
 		]
 	})");
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 	auto typeConfig = std::make_unique<ctypesparser::TypeConfig>();
 	auto demangler = DemanglerProvider::addDemangler(
@@ -462,7 +468,7 @@ TEST_F(ParamReturnTests, x86ExternalCallBasicFunctionality)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 32,
 			"endian" : "little",
@@ -485,6 +491,7 @@ TEST_F(ParamReturnTests, x86ExternalCallBasicFunctionality)
 			}
 		]
 	})");
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 	auto typeConfig = std::make_unique<ctypesparser::TypeConfig>();
 	auto demangler = DemanglerProvider::addDemangler(
@@ -533,7 +540,7 @@ TEST_F(ParamReturnTests, x86ExternalCallFixOnMultiplePlaces)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 32,
 			"endian" : "little",
@@ -574,6 +581,7 @@ TEST_F(ParamReturnTests, x86ExternalCallFixOnMultiplePlaces)
 			}
 		]
 	})");
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 	auto typeConfig = std::make_unique<ctypesparser::TypeConfig>();
 	auto demangler = DemanglerProvider::addDemangler(
@@ -625,7 +633,7 @@ TEST_F(ParamReturnTests, x86ExternalCallFixOnMultiplePlaces)
 //			ret void
 //		}
 //	)");
-//	auto config = Config::fromJsonString(module.get(), R"({
+//	auto c = config::Config::fromJsonString(R"({
 //		"architecture" : {
 //			"bitSize" : 32,
 //			"endian" : "little",
@@ -648,6 +656,7 @@ TEST_F(ParamReturnTests, x86ExternalCallFixOnMultiplePlaces)
 //			}
 //		]
 //	})");
+//	auto config = Config::fromConfig(module.get(), c);
 //	auto abi = AbiProvider::addAbi(module.get(), &config);
 //
 //	pass.runOnModuleCustom(*module, &config, abi);
@@ -688,7 +697,7 @@ TEST_F(ParamReturnTests, x86ExternalCallFixOnMultiplePlaces)
 //			ret void
 //		}
 //	)");
-//	auto config = Config::fromJsonString(module.get(), R"({
+//	auto c = config::Config::fromJsonString(R"({
 //		"architecture" : {
 //			"bitSize" : 32,
 //			"endian" : "little",
@@ -711,6 +720,7 @@ TEST_F(ParamReturnTests, x86ExternalCallFixOnMultiplePlaces)
 //			}
 //		]
 //	})");
+//	auto config = Config::fromConfig(module.get(), c);
 //	auto abi = AbiProvider::addAbi(module.get(), &config);
 //
 //	pass.runOnModuleCustom(*module, &config, abi);
@@ -756,7 +766,7 @@ TEST_F(ParamReturnTests, x86ExternalCallFixOnMultiplePlaces)
 //			ret void
 //		}
 //	)");
-//	auto config = Config::fromJsonString(module.get(), R"({
+//	auto c = config::Config::fromJsonString(R"({
 //		"architecture" : {
 //			"bitSize" : 32,
 //			"endian" : "little",
@@ -779,6 +789,7 @@ TEST_F(ParamReturnTests, x86ExternalCallFixOnMultiplePlaces)
 //			}
 //		]
 //	})");
+//	auto config = Config::fromConfig(module.get(), c);
 //	auto abi = AbiProvider::addAbi(module.get(), &config);
 //
 //	pass.runOnModuleCustom(*module, &config, abi);
@@ -821,7 +832,7 @@ TEST_F(ParamReturnTests, x86ExternalCallFixOnMultiplePlaces)
 //			ret void
 //		}
 //	)");
-//	auto config = Config::fromJsonString(module.get(), R"({
+//	auto c = config::Config::fromJsonString(R"({
 //		"architecture" : {
 //			"bitSize" : 32,
 //			"endian" : "little",
@@ -847,6 +858,7 @@ TEST_F(ParamReturnTests, x86ExternalCallFixOnMultiplePlaces)
 //			}
 //		]
 //	})");
+// 	auto config = Config::fromConfig(module.get(), c);
 //	auto abi = AbiProvider::addAbi(module.get(), &config);
 //
 //	pass.runOnModuleCustom(*module, &config, abi);
@@ -884,7 +896,7 @@ TEST_F(ParamReturnTests, x86ExternalCallFixOnMultiplePlaces)
 //			ret void
 //		}
 //	)");
-//	auto config = Config::fromJsonString(module.get(), R"({
+//	auto c = config::Config::fromJsonString(R"({
 //		"architecture" : {
 //			"bitSize" : 32,
 //			"endian" : "little",
@@ -907,6 +919,7 @@ TEST_F(ParamReturnTests, x86ExternalCallFixOnMultiplePlaces)
 //			}
 //		]
 //	})");
+//	auto config = Config::fromConfig(module.get(), c);
 //	auto abi = AbiProvider::addAbi(module.get(), &config);
 //
 //	pass.runOnModuleCustom(*module, &config, abi);
@@ -946,13 +959,14 @@ TEST_F(ParamReturnTests, x86_64PtrCallBasicFunctionality)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 64,
 			"endian" : "little",
 			"name" : "x86"
 		}
 	})");
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 
 	abi->addRegister(X86_REG_RDI, getGlobalByName("rdi"));
@@ -1014,14 +1028,14 @@ TEST_F(ParamReturnTests, x86_64PtrCallPrevBbIsUsedOnlyIfItIsASinglePredecessor)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 64,
 			"endian" : "little",
 			"name" : "x86"
 		}
 	})");
-
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 
 	abi->addRegister(X86_REG_RDI, getGlobalByName("rdi"));
@@ -1097,7 +1111,7 @@ TEST_F(ParamReturnTests, x86_64ExternalCallUseStacksIf6RegistersUsed)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 64,
 			"endian" : "little",
@@ -1121,6 +1135,7 @@ TEST_F(ParamReturnTests, x86_64ExternalCallUseStacksIf6RegistersUsed)
 			}
 		]
 	})");
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 
 	abi->addRegister(X86_REG_RAX, getGlobalByName("rax"));
@@ -1204,14 +1219,14 @@ TEST_F(ParamReturnTests, x86_64ExternalCallUsesFPRegistersBasic)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 64,
 			"endian" : "little",
 			"name" : "x86"
 		}
 	})");
-
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 
 	abi->addRegister(X86_REG_RAX, getGlobalByName("rax"));
@@ -1285,14 +1300,14 @@ TEST_F(ParamReturnTests, x86_64ExternalCallUsesFPRegisters)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 64,
 			"endian" : "little",
 			"name" : "x86"
 		}
 	})");
-
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 
 	abi->addRegister(X86_REG_RAX, getGlobalByName("rax"));
@@ -1382,13 +1397,14 @@ TEST_F(ParamReturnTests, x86_64UsesJustContinuousSequenceOfRegisters)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 64,
 			"endian" : "little",
 			"name" : "x86"
 		}
 	})");
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 
 	abi->addRegister(X86_REG_RAX, getGlobalByName("rax"));
@@ -1451,7 +1467,7 @@ TEST_F(ParamReturnTests, ms_x64PtrCallBasicFunctionality)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 64,
 			"endian" : "little",
@@ -1463,6 +1479,7 @@ TEST_F(ParamReturnTests, ms_x64PtrCallBasicFunctionality)
 			{"name" : "gcc"}
 		]
 	})");
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 
 	abi->addRegister(X86_REG_RCX, getGlobalByName("rcx"));
@@ -1524,7 +1541,7 @@ TEST_F(ParamReturnTests, ms_x64PtrCallPrevBbIsUsedOnlyIfItIsASinglePredecessor)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 64,
 			"endian" : "little",
@@ -1536,7 +1553,7 @@ TEST_F(ParamReturnTests, ms_x64PtrCallPrevBbIsUsedOnlyIfItIsASinglePredecessor)
 			{"name" : "gcc"}
 		]
 	})");
-
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 
 	abi->addRegister(X86_REG_RCX, getGlobalByName("rcx"));
@@ -1608,7 +1625,7 @@ TEST_F(ParamReturnTests, ms_x64ExternalCallUseStacksIf4RegistersUsed)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 64,
 			"endian" : "little",
@@ -1638,6 +1655,7 @@ TEST_F(ParamReturnTests, ms_x64ExternalCallUseStacksIf4RegistersUsed)
 		]
 
 	})");
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 
 	abi->addRegister(X86_REG_RAX, getGlobalByName("rax"));
@@ -1715,7 +1733,7 @@ TEST_F(ParamReturnTests, ms_x64ExternalCallUsesFPRegisters)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 64,
 			"endian" : "little",
@@ -1727,7 +1745,7 @@ TEST_F(ParamReturnTests, ms_x64ExternalCallUsesFPRegisters)
 			{"name" : "gcc"}
 		]
 	})");
-
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 
 	abi->addRegister(X86_REG_RAX, getGlobalByName("rax"));
@@ -1794,7 +1812,7 @@ TEST_F(ParamReturnTests, ms_x64UsesJustContinuousSequenceOfRegisters)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 64,
 			"endian" : "little",
@@ -1806,6 +1824,7 @@ TEST_F(ParamReturnTests, ms_x64UsesJustContinuousSequenceOfRegisters)
 			{"name" : "gcc"}
 		]
 	})");
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 
 	abi->addRegister(X86_REG_RAX, getGlobalByName("rax"));
@@ -1872,7 +1891,7 @@ TEST_F(ParamReturnTests, ms_x64ExternalCallUsesFPRegistersAdvanced)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 64,
 			"endian" : "little",
@@ -1884,7 +1903,7 @@ TEST_F(ParamReturnTests, ms_x64ExternalCallUsesFPRegistersAdvanced)
 			{"name" : "gcc"}
 		]
 	})");
-
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 
 	abi->addRegister(X86_REG_RAX, getGlobalByName("rax"));
@@ -1954,7 +1973,7 @@ TEST_F(ParamReturnTests, ms_x64ExternalCallUsesFPRegistersAdvanced)
 //			ret void
 //		}
 //	)");
-//	auto config = Config::fromJsonString(module.get(), R"({
+//	auto c = config::Config::fromJsonString(R"({
 //		"architecture" : {
 //			"bitSize" : 32,
 //			"endian" : "little",
@@ -1985,6 +2004,7 @@ TEST_F(ParamReturnTests, ms_x64ExternalCallUsesFPRegistersAdvanced)
 //			}
 //		]
 //	})");
+//	auto config = Config::fromConfig(module.get(), c);
 //	auto abi = AbiProvider::addAbi(module.get(), &config);
 //
 
@@ -2002,13 +2022,14 @@ TEST_F(ParamReturnTests, ppcPtrCallBasicFunctionality)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 32,
 			"endian" : "big",
 			"name" : "powerpc"
 		}
 	})");
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 
 	abi->addRegister(PPC_REG_R3, getGlobalByName("r3"));
@@ -2057,13 +2078,14 @@ TEST_F(ParamReturnTests, ppcExternalCallBasicFunctionality)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 32,
 			"endian" : "big",
 			"name" : "powerpc"
 		}
 	})");
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 
 	abi->addRegister(PPC_REG_R3, getGlobalByName("r3"));
@@ -2117,13 +2139,14 @@ TEST_F(ParamReturnTests, ppcExternalCallBasicFPFunctionality)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 32,
 			"endian" : "big",
 			"name" : "powerpc"
 		}
 	})");
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 
 	abi->addRegister(PPC_REG_R3, getGlobalByName("r3"));
@@ -2179,13 +2202,14 @@ TEST_F(ParamReturnTests, ppcExternalCallDoNotUseObjectsIfTheyAreNotRegisters)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 32,
 			"endian" : "big",
 			"name" : "powerpc"
 		}
 	})");
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 
 	auto typeConfig = std::make_unique<ctypesparser::TypeConfig>();
@@ -2228,13 +2252,14 @@ TEST_F(ParamReturnTests, ppcExternalCallFilterRegistersOnMultiplePlaces)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 32,
 			"endian" : "big",
 			"name" : "powerpc"
 		}
 	})");
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 	abi->addRegister(PPC_REG_R3, getGlobalByName("r3"));
 	abi->addRegister(PPC_REG_R4, getGlobalByName("r4"));
@@ -2293,13 +2318,14 @@ TEST_F(ParamReturnTests, ppcExternalCallDoNotUseAllRegisters)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 32,
 			"endian" : "big",
 			"name" : "powerpc"
 		}
 	})");
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 
 	abi->addRegister(PPC_REG_R1, getGlobalByName("r1"));
@@ -2349,13 +2375,14 @@ TEST_F(ParamReturnTests, ppcExternalCallSortRegistersIntoCorrectOrder)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 32,
 			"endian" : "big",
 			"name" : "powerpc"
 		}
 	})");
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 
 	abi->addRegister(PPC_REG_R3, getGlobalByName("r3"));
@@ -2410,7 +2437,7 @@ TEST_F(ParamReturnTests, ppcExternalCallDoNotUseStacksIfLessThan7RegistersUsed)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 32,
 			"endian" : "big",
@@ -2429,6 +2456,7 @@ TEST_F(ParamReturnTests, ppcExternalCallDoNotUseStacksIfLessThan7RegistersUsed)
 			}
 		]
 	})");
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 
 	abi->addRegister(PPC_REG_R3, getGlobalByName("r3"));
@@ -2479,14 +2507,14 @@ TEST_F(ParamReturnTests, ppc64PtrCallBasicFunctionality)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 64,
 			"endian" : "big",
 			"name" : "powerpc64"
 		}
 	})");
-
+	auto config = Config::fromConfig(module.get(), c);
 	AbiPowerpc64 abi(module.get(), &config);
 
 	abi.addRegister(PPC_REG_R3, getGlobalByName("r3"));
@@ -2563,13 +2591,14 @@ TEST_F(ParamReturnTests, armPtrCallBasicFunctionality)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 32,
 			"endian" : "little",
 			"name" : "arm"
 		}
 	})");
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 
 	abi->addRegister(ARM_REG_R0, getGlobalByName("r0"));
@@ -2618,13 +2647,14 @@ TEST_F(ParamReturnTests, armExternalCallBasicFunctionality)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 32,
 			"endian" : "little",
 			"name" : "arm"
 		}
 	})");
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 
 	abi->addRegister(ARM_REG_R0, getGlobalByName("r0"));
@@ -2686,7 +2716,7 @@ TEST_F(ParamReturnTests, armExternalCallUseStacksIf4RegistersUsed)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 32,
 			"endian" : "little",
@@ -2709,6 +2739,7 @@ TEST_F(ParamReturnTests, armExternalCallUseStacksIf4RegistersUsed)
 			}
 		]
 	})");
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 
 	abi->addRegister(ARM_REG_R0, getGlobalByName("r0"));
@@ -2779,13 +2810,14 @@ TEST_F(ParamReturnTests, arm64PtrCallBasicFunctionality)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 64,
 			"endian" : "little",
 			"name" : "arm aarch64"
 		}
 	})");
+	auto config = Config::fromConfig(module.get(), c);
 	AbiArm64 abi(module.get(), &config);
 
 	abi.addRegister(ARM64_REG_X0, getGlobalByName("x0"));
@@ -2836,14 +2868,14 @@ TEST_F(ParamReturnTests, arm64ExternalCallBasicFunctionality)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 64,
 			"endian" : "little",
 			"name" : "arm aarch64"
 		}
 	})");
-
+	auto config = Config::fromConfig(module.get(), c);
 	AbiArm64 abi(module.get(), &config);
 
 	abi.addRegister(ARM64_REG_X0, getGlobalByName("x0"));
@@ -2913,7 +2945,7 @@ TEST_F(ParamReturnTests, arm64ExternalCallUseStacksIf8RegistersUsed)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 64,
 			"endian" : "little",
@@ -2936,6 +2968,7 @@ TEST_F(ParamReturnTests, arm64ExternalCallUseStacksIf8RegistersUsed)
 			}
 		]
 	})");
+	auto config = Config::fromConfig(module.get(), c);
 	AbiArm64 abi(module.get(), &config);
 
 	abi.addRegister(ARM64_REG_X0, getGlobalByName("x0"));
@@ -3023,13 +3056,14 @@ TEST_F(ParamReturnTests, arm64ExternalCallHasDouleParameter)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 64,
 			"endian" : "little",
 			"name" : "arm aarch64"
 		}
 	})");
+	auto config = Config::fromConfig(module.get(), c);
 	AbiArm64 abi(module.get(), &config);
 
 	abi.addRegister(ARM64_REG_X0, getGlobalByName("x0"));
@@ -3103,7 +3137,7 @@ TEST_F(ParamReturnTests, arm64ExternalCallHasDouleParameter)
 //			ret void
 //		}
 //	)");
-//	auto config = Config::fromJsonString(module.get(), R"({
+//	auto c = config::Config::fromJsonString(R"({
 //		"architecture" : {
 //			"bitSize" : 32,
 //			"endian" : "big",
@@ -3168,6 +3202,7 @@ TEST_F(ParamReturnTests, arm64ExternalCallHasDouleParameter)
 //			}
 //		]
 //	})");
+//	auto config = Config::fromConfig(module.get(), c);
 //	auto abi = AbiProvider::addAbi(module.get(), &config);
 //
 
@@ -3185,13 +3220,14 @@ TEST_F(ParamReturnTests, mipsPtrCallBasicFunctionality)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 32,
 			"endian" : "little",
 			"name" : "mips"
 		}
 	})");
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 
 	abi->addRegister(MIPS_REG_A0, getGlobalByName("a0"));
@@ -3236,13 +3272,14 @@ TEST_F(ParamReturnTests, mipsExternalCallBasicFunctionality)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 32,
 			"endian" : "little",
 			"name" : "mips"
 		}
 	})");
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 
 	abi->addRegister(MIPS_REG_A0, getGlobalByName("a0"));
@@ -3296,7 +3333,7 @@ TEST_F(ParamReturnTests, mipsExternalCallUseStacksIf4RegistersUsed)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 32,
 			"endian" : "little",
@@ -3319,6 +3356,7 @@ TEST_F(ParamReturnTests, mipsExternalCallUseStacksIf4RegistersUsed)
 			}
 		]
 	})");
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 
 	abi->addRegister(MIPS_REG_A0, getGlobalByName("a0"));
@@ -3382,14 +3420,14 @@ TEST_F(ParamReturnTests, mips64PtrCallBasicFunctionality)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 64,
 			"endian" : "little",
 			"name" : "mips64"
 		}
 	})");
-
+	auto config = Config::fromConfig(module.get(), c);
 	AbiMips64 abi(module.get(), &config);
 
 	abi.addRegister(MIPS_REG_A0, getGlobalByName("a0"));
@@ -3438,14 +3476,14 @@ TEST_F(ParamReturnTests, mips64ExternalCallBasicFunctionality)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 64,
 			"endian" : "little",
 			"name" : "mips64"
 		}
 	})");
-
+	auto config = Config::fromConfig(module.get(), c);
 	AbiMips64 abi(module.get(), &config);
 
 	abi.addRegister(MIPS_REG_A0, getGlobalByName("a0"));
@@ -3511,7 +3549,7 @@ TEST_F(ParamReturnTests, mips64ExternalCallUseStacksIf8RegistersUsed)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 64,
 			"endian" : "little",
@@ -3534,7 +3572,7 @@ TEST_F(ParamReturnTests, mips64ExternalCallUseStacksIf8RegistersUsed)
 			}
 		]
 	})");
-
+	auto config = Config::fromConfig(module.get(), c);
 	AbiMips64 abi(module.get(), &config);
 
 	abi.addRegister(MIPS_REG_A0, getGlobalByName("a0"));
@@ -3622,7 +3660,7 @@ TEST_F(ParamReturnTests, x86FastcallBasic)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 32,
 			"endian" : "little",
@@ -3654,7 +3692,7 @@ TEST_F(ParamReturnTests, x86FastcallBasic)
 			}
 		]
 	})");
-
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 	abi->addRegister(X86_REG_EAX, getGlobalByName("eax"));
 	abi->addRegister(X86_REG_ECX, getGlobalByName("ecx"));
@@ -3718,7 +3756,7 @@ TEST_F(ParamReturnTests, x86FastcallLargeTypeCatch)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 32,
 			"endian" : "little",
@@ -3752,7 +3790,7 @@ TEST_F(ParamReturnTests, x86FastcallLargeTypeCatch)
 		]
 
 	})");
-
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 	abi->addRegister(X86_REG_ECX, getGlobalByName("ecx"));
 	abi->addRegister(X86_REG_EAX, getGlobalByName("eax"));
@@ -3809,7 +3847,7 @@ TEST_F(ParamReturnTests, x86PascalBasic)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 32,
 			"endian" : "little",
@@ -3836,7 +3874,7 @@ TEST_F(ParamReturnTests, x86PascalBasic)
 			}
 		]
 	})");
-
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 
 	auto typeConfig = std::make_unique<ctypesparser::TypeConfig>();
@@ -3887,7 +3925,7 @@ TEST_F(ParamReturnTests, x86PascalFastcallBasic)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 32,
 			"endian" : "little",
@@ -3919,7 +3957,7 @@ TEST_F(ParamReturnTests, x86PascalFastcallBasic)
 			}
 		]
 	})");
-
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 	abi->addRegister(X86_REG_EAX, getGlobalByName("eax"));
 	abi->addRegister(X86_REG_EDX, getGlobalByName("edx"));
@@ -3987,7 +4025,7 @@ TEST_F(ParamReturnTests, x86PascalFastcallLargeType)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 32,
 			"endian" : "little",
@@ -4019,7 +4057,7 @@ TEST_F(ParamReturnTests, x86PascalFastcallLargeType)
 			}
 		]
 	})");
-
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 	abi->addRegister(X86_REG_EAX, getGlobalByName("eax"));
 	abi->addRegister(X86_REG_EDX, getGlobalByName("edx"));
@@ -4085,7 +4123,7 @@ TEST_F(ParamReturnTests, x86WatcomBasic)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 32,
 			"endian" : "little",
@@ -4112,7 +4150,7 @@ TEST_F(ParamReturnTests, x86WatcomBasic)
 			}
 		]
 	})");
-
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 	abi->addRegister(X86_REG_EAX, getGlobalByName("eax"));
 	abi->addRegister(X86_REG_EBX, getGlobalByName("ebx"));
@@ -4179,7 +4217,7 @@ TEST_F(ParamReturnTests, x86WatcomPassDouble)
 			ret void
 		}
 	)");
-	auto config = Config::fromJsonString(module.get(), R"({
+	auto c = config::Config::fromJsonString(R"({
 		"architecture" : {
 			"bitSize" : 32,
 			"endian" : "little",
@@ -4206,7 +4244,7 @@ TEST_F(ParamReturnTests, x86WatcomPassDouble)
 			}
 		]
 	})");
-
+	auto config = Config::fromConfig(module.get(), c);
 	auto abi = AbiProvider::addAbi(module.get(), &config);
 	abi->addRegister(X86_REG_EAX, getGlobalByName("eax"));
 	abi->addRegister(X86_REG_EDX, getGlobalByName("edx"));
