+++ tests/bin2llvmir/optimizations/unreachable_funcs/unreachable_funcs_tests.cpp
@@ -61,7 +61,7 @@ TEST_F(UnreachableFuncsTests, addressOfFunc01)
 		  ret i32 0
 		}
 	)");
-	ConfigProvider::addConfigJsonString(module.get(), R"({
+	auto config = config::Config::fromJsonString(R"({
 		"mainAddress" : "0x1000",
 		"functions" :
 		[
@@ -71,6 +71,7 @@ TEST_F(UnreachableFuncsTests, addressOfFunc01)
 			}
 		]
 	})");
+	ConfigProvider::addConfig(module.get(), config);
 
 	runOnModule();
 
@@ -121,7 +122,7 @@ TEST_F(UnreachableFuncsTests, funcInGlobalVarInit)
 		  ret i32 0
 		}
 	)");
-	ConfigProvider::addConfigJsonString(module.get(), R"({
+	auto config = config::Config::fromJsonString(R"({
 		"mainAddress" : "0x1000",
 		"functions" :
 		[
@@ -131,6 +132,7 @@ TEST_F(UnreachableFuncsTests, funcInGlobalVarInit)
 			}
 		]
 	})");
+	ConfigProvider::addConfig(module.get(), config);
 
 	runOnModule();
 
@@ -241,7 +243,7 @@ TEST_F(UnreachableFuncsTests, indirectCall01)
 		  ret i32 0
 		}
 	)");
-	ConfigProvider::addConfigJsonString(module.get(), R"({
+	auto config = config::Config::fromJsonString(R"({
 		"mainAddress" : "0x1000",
 		"functions" :
 		[
@@ -251,6 +253,7 @@ TEST_F(UnreachableFuncsTests, indirectCall01)
 			}
 		]
 	})");
+	ConfigProvider::addConfig(module.get(), config);
 
 	runOnModule();
 
@@ -337,7 +340,7 @@ TEST_F(UnreachableFuncsTests, mainAndFuncs01)
 		  ret i32 0
 		}
 	)");
-	ConfigProvider::addConfigJsonString(module.get(), R"({
+	auto config = config::Config::fromJsonString(R"({
 		"mainAddress" : "0x1000",
 		"functions" :
 		[
@@ -347,6 +350,7 @@ TEST_F(UnreachableFuncsTests, mainAndFuncs01)
 			}
 		]
 	})");
+	ConfigProvider::addConfig(module.get(), config);
 
 	runOnModule();
 
@@ -395,7 +399,7 @@ TEST_F(UnreachableFuncsTests, mainAndFuncs02)
 		  ret i32 0
 		}
 	)");
-	ConfigProvider::addConfigJsonString(module.get(), R"({
+	auto config = config::Config::fromJsonString(R"({
 		"mainAddress" : "0x1000",
 		"functions" :
 		[
@@ -405,6 +409,7 @@ TEST_F(UnreachableFuncsTests, mainAndFuncs02)
 			}
 		]
 	})");
+	ConfigProvider::addConfig(module.get(), config);
 
 	runOnModule();
 
@@ -455,7 +460,7 @@ TEST_F(UnreachableFuncsTests, mainAndFuncs03)
 		  ret i32 0
 		}
 	)");
-	ConfigProvider::addConfigJsonString(module.get(), R"({
+	auto config = config::Config::fromJsonString(R"({
 		"mainAddress" : "0x1000",
 		"functions" :
 		[
@@ -465,6 +470,7 @@ TEST_F(UnreachableFuncsTests, mainAndFuncs03)
 			}
 		]
 	})");
+	ConfigProvider::addConfig(module.get(), config);
 
 	runOnModule();
 
@@ -530,7 +536,7 @@ TEST_F(UnreachableFuncsTests, mainAndFuncs04)
 		  ret i32 0
 		}
 	)");
-	ConfigProvider::addConfigJsonString(module.get(), R"({
+	auto config = config::Config::fromJsonString(R"({
 		"mainAddress" : "0x1000",
 		"functions" :
 		[
@@ -540,6 +546,7 @@ TEST_F(UnreachableFuncsTests, mainAndFuncs04)
 			}
 		]
 	})");
+	ConfigProvider::addConfig(module.get(), config);
 
 	runOnModule();
 
@@ -583,7 +590,7 @@ TEST_F(UnreachableFuncsTests, mainOnlyDeclaration)
 		; Can't be optimized.
 		declare i32 @main(i32 %arg1, i8** nocapture %arg2)
 	)");
-	ConfigProvider::addConfigJsonString(module.get(), R"({
+	auto config = config::Config::fromJsonString(R"({
 		"mainAddress" : "0x1000",
 		"functions" :
 		[
@@ -593,6 +600,7 @@ TEST_F(UnreachableFuncsTests, mainOnlyDeclaration)
 			}
 		]
 	})");
+	ConfigProvider::addConfig(module.get(), config);
 
 	runOnModule();
 
@@ -684,7 +692,7 @@ TEST_F(UnreachableFuncsTests, onlyMain01)
 		  ret i32 0
 		}
 	)");
-	ConfigProvider::addConfigJsonString(module.get(), R"({
+	auto config = config::Config::fromJsonString(R"({
 		"mainAddress" : "0x1000",
 		"functions" :
 		[
@@ -694,6 +702,7 @@ TEST_F(UnreachableFuncsTests, onlyMain01)
 			}
 		]
 	})");
+	ConfigProvider::addConfig(module.get(), config);
 
 	runOnModule();
 
@@ -729,7 +738,7 @@ TEST_F(UnreachableFuncsTests, onlyMain02)
 		  ret i32 0
 		}
 	)");
-	ConfigProvider::addConfigJsonString(module.get(), R"({
+	auto config = config::Config::fromJsonString(R"({
 		"mainAddress" : "0x1000",
 		"functions" :
 		[
@@ -739,6 +748,7 @@ TEST_F(UnreachableFuncsTests, onlyMain02)
 			}
 		]
 	})");
+	ConfigProvider::addConfig(module.get(), config);
 
 	runOnModule();
